use real box

+++ src/app/box.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { Box } from './models/Box';
 import { StorageService } from './storage.service';
 import { Guid } from 'guid-typescript';
+import { Vocabulary } from './models/Vocabulary';
 
 @Injectable({
   providedIn: 'root',
@@ -42,6 +43,33 @@ export class BoxService {
     await this.storageService.save(this.STORAGE_KEY, boxes);
   }
 
+  async insertOrUpdateVocabulary(boxId: string, vocabulary: Vocabulary) {
+    vocabulary.id =
+      vocabulary.id?.length > 0 ? vocabulary.id : Guid.create().toString();
+
+    var boxes = await this.getAll();
+    if (!boxes) {
+      console.info('There are no boxes stored');
+      return;
+    }
+
+    const stored = boxes.find((x) => x.id === boxId);
+    if (!stored) {
+      console.info('could not find box with id:', boxId);
+      return;
+    }
+
+    stored.vocabularies = stored.vocabularies?.filter(
+      (x) => x.id !== vocabulary.id
+    );
+
+    stored.vocabularies = stored.vocabularies ? stored.vocabularies : [];
+    stored.vocabularies.push(vocabulary);
+
+    await this.storageService.delete(this.STORAGE_KEY);
+    await this.storageService.save(this.STORAGE_KEY, boxes);
+  }
+
   async delete(id: string): Promise<any> {
     var boxes = await this.getAll();
     if (!boxes) {
@@ -49,6 +77,27 @@ export class BoxService {
     }
 
     boxes = boxes.filter((x) => x.id !== id);
+    await this.storageService.delete(this.STORAGE_KEY);
+    await this.storageService.save(this.STORAGE_KEY, boxes);
+  }
+
+  async deleteVocabulary(boxId: string, vocabularyId: string) {
+    var boxes = await this.getAll();
+    if (!boxes) {
+      console.info('There are no boxes stored');
+      return;
+    }
+
+    const stored = boxes.find((x) => x.id === boxId);
+    if (!stored) {
+      console.info('could not find box with id:', boxId);
+      return;
+    }
+
+    stored.vocabularies = stored.vocabularies?.filter(
+      (x) => x.id !== vocabularyId
+    );
+
     await this.storageService.delete(this.STORAGE_KEY);
     await this.storageService.save(this.STORAGE_KEY, boxes);
   }
